Incredible lander module
the liquid fuel engines on this thing made it from launch to halfway out of the atmosphere on Kerbin and the parachutes successfully landed it with the gear extended.

This thing is not meant for kerbin escape so it should be more than sufficient to get back into Minmus orbit...etc.