Built & launched v-02 craft
delightful new science probe. Small & unmanned, this probe contains a few basic science components (thermometer, Goo, & Gravioli detector), an antenna for transmission, and is self-sufficient for power.

The first stage gets you to upper atmo & the 2nd stage is good for orbital maneuver with 1/3 of the tank remaining afterwards. I meant the fuel tank to be discarded after orbit established but had so much fuel left that decided to keep it. Had to rearrange the staging for the decouplers to only pop the clamshell fairing and not detach anything.