Launched probes, got science!
Launched a few more unmanned probes & transmitted data back for the science wins. Launched one big lander up that is manned & has a bit of fuel in the 2nd stage rocket & a small but full tank with poodle engine still untouched. Can land back on kerbin if wanted. Building all of these with the intention of re-entry & safe landing, but accidentally making the rockets capable of orbit maneuvers. =/ oh well.