2 new rockets saved, 1 new satellite in orbit
The Condor Substation doesn't launch yet. need to add more thrust to keep it stable. Idea was to launch an unmanned satellite with empty crew capsule + docking port. Then to be able to dock with it to unload a scientist at some point. The Atomic engine on it should provide plenty of thrust to do some orbit adjustments also.