Excellent 3 stage launcher. looks like Taj Mahal.
Uploaded screenshot. =)
It's a great launcher that can transmit data and get into orbit with semi-reasonable remaining fuel. Takes a bit of skill to launch because it gets a wobble but the remaining fuel amount is enough to make up for maneuvering to fix orbit once you get out there.

This could be useful for combining 2-3 of the same thing in orbit and shipping them out somewhere on the end of a big engine. each module can have a different science thing attached and it'd be manned by at least several kerbins (we can ferry the rest off once they are attached but the pods have reaction wheels and monopropellant which are useful parts.